requirements.txt
albumentations
cvzone
matplotlib
mediapipe
numpy
opencv-python 
pandas
Pillow
pyttsx3
pytz
scikit-learn
timm
tqdm
wandb